Fixed transaction form for adding and editing transactions
Added haptic feedback when a checkbox is tapped
Found solution to the bug where a change to a transaction would return the user to the forecast list view
Changed from using Float to Double for currency values

+++ finance-forecast/Models/Forecast.swift
@@ -5,12 +5,12 @@ struct Forecast: Hashable, Identifiable {
     var month: Int
     var year: Int
     var transactions: [Transaction]
-    var endBalance: Float = 0
+    var endBalance: Double = 0
     
     let monthFormatter = DateFormatter()
     
-    func transactionTotal() -> Float {
-        var total: Float = 0
+    func transactionTotal() -> Double {
+        var total: Double = 0
         for tx in self.transactions {
             total += tx.amount
         }
